refactor: added useSearchParameters hook and renamed useViewParams to useViewParameters
