+++ src/Components/Main/Pages/NotFoundPage/NotFoundPage.tsx
@@ -1,10 +1,15 @@
 import type { FC, ReactElement } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Error } from '../../../Common/Error/Error';
+import { useTypedSelector } from '../../../../hooks/useTypedSelector';
+import { viewSelector } from '../../../../redux/selectors/selectors';
+import { errorText } from '../../../../utils/languageLocalization/errorText';
 
 export const NotFoundPage: FC = (): ReactElement => {
+  const { languageMode } = useTypedSelector(viewSelector);
+
   const { pathname } = useLocation();
-  const message = `The requested URL "${pathname}" was not found on this server!`;
+  const message = `${errorText.notFoundPt1[languageMode]} "${pathname}" ${errorText.notFoundPt2[languageMode]}!`;
 
   return <Error code={404} message={message} />;
 };
